Eliminate some unnecessary boilerplate for applications.
Related: #12
