+++ blivet/tasks/task.py
@@ -23,6 +23,8 @@ import abc
 
 from six import add_metaclass
 
+from . import availability
+
 @add_metaclass(abc.ABCMeta)
 class Task(object):
     """ An abstract class that represents some task. """
@@ -101,3 +103,38 @@ class UnimplementedTask(Task):
 
     def doTask(self, *args, **kwargs):
         raise NotImplementedError()
+
+@add_metaclass(abc.ABCMeta)
+class BasicApplication(Task):
+    """ A task representing an application. """
+
+    app_name = abc.abstractproperty(doc="The name of the application.")
+
+    @classmethod
+    def _app(cls):
+        return availability.application(cls.app_name)
+
+    # TASK methods
+
+    @classmethod
+    def available(cls):
+        return cls._app().available
+
+    @property
+    def _unavailable(self):
+        if not self._app().available:
+            return "application %s is not available" % self._app()
+
+        return False
+
+    @property
+    def dependsOn(self):
+        return []
+
+    @property
+    def unable(self):
+        return False
+
+    @property
+    def unready(self):
+        return False
